remove conv directory and put the small amount of code in the appropriate places
