--- conv/errors.go
@@ -1,7 +1,0 @@
-package conv
-
-import "errors"
-
-var (
-	errInvalidAddress = errors.New("invalid address format, expected [protocol://][<NODE_ID>@]<IPv4>:<PORT>")
-)
